Deal with images hosted in Cloudinary
A helper tag to generate image tags with consistent resize settings.
This means if later I decide to use larger images due to a template
change or something, I can change all the images in one go from this
plugin.

The 600x600 constraint was fairly arbitrary, so this should make future
changes easier.

I also made portrait images that dont fill the width of the page to be
centered.