Don't just strip newlines, keep a space there
The index page would have content like this:

> My flight from Melbourne airport was delayed, but luckily my stop over
> inAuckland was really long.While

Note the "inAuckland" and "long.While".

This is because Markdown normally doesn't mind newlines in the middle of
content because you need two spaces at the end of the line to result in
a proper line break. This means the source of my posts can be limited to
80 character line lengths.

However the liquid filter of `strip_newlines` seems to literally strip
them out and not replace them with a space, which doesn't seem to make
much sense in my case.

Using `newlines_to_br` instead seems to behave better, but doesn't
actually seem to put `<br />` tags in, not really sure why not
considering the HTML is supposedly stripped before this filter, but
either way the end result is better:

> My flight from Melbourne airport was delayed, but luckily my stop over
> in Auckland was really long. While